requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.8.1
bleach==6.1.0
coverage==7.6.10
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-crispy-forms==2.2
django-parler==2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
django-summernote==0.8.20.0
djangorestframework==3.15.2
easy-thumbnails==2.8.5
pillow==10.4.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
webencodings==0.5.1
